+++ src/components/DirectoryList.js
@@ -1,4 +1,5 @@
 import { setItem, getItem, removeItem } from "../utils/storage.js";
+import { totalDirectoryList } from "../utils/totalDirectoryList.js";
 
 export default function DirectoryList({ $target, initialState }) {
   const $directoryList = document.createElement("div");
@@ -9,37 +10,6 @@ export default function DirectoryList({ $target, initialState }) {
   this.setState = (nextState) => {
     this.state = nextState;
     this.render();
-  };
-
-  const totalDirectoryList = (directorys) => {
-    return `
-      <ul>
-        ${directorys
-          .map((directory) => {
-            const listToggleState = `isOpened-${directory.id}`;
-            const display = getItem(listToggleState) || "none";
-            return `
-              <li data-id="${directory.id}" name="list">
-              <span class="list-title"><button class="toggle-button" isClosed="true">></button>${
-                directory.title
-              }</span>
-              <span class="list-button"><button name="add-button" data-id="${
-                directory.id
-              }">+</button>
-              <button name="remove-button" data-id=${directory.id}>-</button></span>
-                ${
-                  directory.documents.length
-                    ? `<ul style="display: ${display}">${totalDirectoryList(
-                        directory.documents
-                      )}</ul>`
-                    : `<ul style="display: ${display}"><li>하위 페이지가 없습니다.</li></ul>`
-                }
-              </li>
-            `;
-          })
-          .join("")}
-      </ul>
-    `;
   };
 
   this.render = () => {
